Add files via upload
prompt: Write a program which will ask the user to enter a collection of numbers at the keyboard, and then analyse those numbers. The 'analysis' will comprise finding the max and the min of the collection of numbers, as well as their mean*. Notice that there's no need to store all the numbers being entered. (Please don't simply use the inbuilt Python functions for these computations -- you should use this question to gain practice in performing simple calculations directly.)