Create HW2Q1
Prompt: Continue development of the matrix program we've been working on in lecture. You should be building a program which has functions to add, subtract, and multiply matrices. Your program should also be able to compute the transpose of a matrix, and determine if a matrix is symmetric. You should also handle the special cases of multiplying a vector by a matrix, computing an inner product (aka scalar product, aka dot product) and computing a weighted inner product (all this will be described in class). It would be good to have a reasonable amount of 'idiot-proofing' and to make it easy for a user to enter data either from the keyboard or from a file. If you'd like to try exploring making a formal class to do all this, that would not be inappropriate, but is not required this time.