Create HW2Q2
Imagine that points in the plane have two coordinates (not necessarily integer-valued). Write a program which will ask the user to enter several points (ie a pair of numbers per line) and output three lists, one being those tuples corresponding to points above the line y=x, one being those on that line, and the other being those points below that line. Your program should also decide which of those three tuples has the most points, and which the least, and print to the screen (for example) "Above > Below > On" (if there are more points above the line, than below the line, than on the line). It would be good to allow the user to specify a file from which your program can read data. You might also consider extensions to this program which allow counting of points on either side of other geometric shapes.