Add files via upload
Prompt:

Build a model of a bus company.

There should be at least 12 destinations (don't have too many!), and several buses.
Each bus should be built by a Bus class, and should have a chosen number of seats, and be associated with a pair of destinations (one to leave from and one to arrive at -- these might get updated after each trip), and people (to choose the bus route they want and to sit in the buses). The buses should indicate if they're full, but until then should allow people to join the bus.
You should have 'towns' (as destinations) which would also be built by a Town class (these needn't have many attributes -- enough so that they can act as destinations and so that they can be queried to see which people and buses are there). Each town should have a 'road' to at least two other towns but not directly to every other town!!), though this road need only be a pointer to that other town (it needn't have its own class or content or name).
Buses, of course, will only be able to travel on 'roads' (so this will constrain possible routes).
Build such other structure to your model as makes sense to you without becoming overly sophisticated!!!
Your system should allow for file backup in the sense that you can save the state of the system when you like (to a text file), quit the program, and then restart the program and call a method to read the (backup) file to repopulate the data to recover the state of the system.


-- I have not figured out how to take information from the file and save it(specifically the people objects)