Update HW4
My ReadMe: 
'''I started by creating the List class and the node class because we couldn't use python's list. I worked with
Professor Bailey's code and edited it so I could pass data - which is an object- into the node. I had difficulty
with printing the list, so I created a printList function which went through each element in my list and added
the object to a string, so I could print them all together. I also created multiple find functions so I could find
specific things I needed, like finding an object by its location or name. I then worked with the people objects and
their friends. I was able to add friends but I had some errors with removing friends because of the index out of range,
I noticed that previous and next nodes might not exist so I had to be careful with them. I wrote a borrow Book function
that used regular for loops, but it wasn't efficient so I tried to create a breadth first search function. The BFS was
difficult for me because I had trouble writing code in the correct order and couldn't tell when to call the recursive
method, but after a lot of work, it was successful.
First I created a bunch of functions to borrow books by certain attributes. They only differed by one line so I tried
to find a better solution. I tried to build a bucket sort but it was really hard so I ended up creating the same code
for multiple situations(again, I would like to work on my efficiency)
    My BFS function is used when a person wants to get their book back- that they have annotated- BFS searches for the
    book and returns the person object who has it. I was having an error when returning the person, Austin helped me with
     the return function (my p variable wasn't updating its information). This was because the local variables weren't
    being passed up and down the tree, so the information was being lost.
The sorting was also difficult, I first created the sorting functions in a seperate file but then realized that it would
be better to have it in the list class. (I tried to use bucket list sorting, but it was unsuccesful.) I have a lot
of Merge sorting functions that are basically the same except for one line, if I had more time I would like to
find a more efficient way to sort by properties.
Finally, in my classy action file, I call all my functions. I worked with a set number of people and books throughout
this project so I had little time at the end to allow the user to input an amount of people and books. I started that
but didn't have time to finish it. I would defintely like to figure out how to do that because that has been a problem
I've had a couple times. I wiggled around the user input so I could just use the objects I had already created and
done practice tests with. This was an extremely hard project, and there are some ideas I still want to incorporate, but
overall I am happy about how it played out.'''